+++ app/models/concerns/submittable/solvable.rb
@@ -6,7 +6,7 @@ module Solvable
   end
 
   def run_tests!(params)
-    language.run_tests!(params.merge(locale: locale, expectations: expectations))
+    language.run_tests!(params.merge(locale: locale, expectations: accumulated_expectations))
   end
 end
